remove old driver
