+++ internal/cli/command/install.go
@@ -3,7 +3,7 @@ package command
 // import (
 // 	"context"
 
-// 	"github.com/rocketblend/rocketblend/pkg/driver/reference"
+// 	"github.com/rocketblend/rocketblend/pkg/reference"
 // 	"github.com/spf13/cobra"
 // )
