+++ pkg/configurator/configurator.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/flowshot-io/x/pkg/logger"
 	"github.com/mitchellh/mapstructure"
-	"github.com/rocketblend/rocketblend/pkg/driver/runtime"
+	"github.com/rocketblend/rocketblend/pkg/runtime"
 	"github.com/rocketblend/rocketblend/pkg/types"
 	"github.com/rocketblend/rocketblend/pkg/validator"
 	"github.com/spf13/viper"
