--- pkg/driver/blendfile/scripts/createScript.gopy
@@ -1,3 +1,0 @@
-import bpy
-bpy.ops.wm.save_as_mainfile(filepath=r'{{ .path }}')
-bpy.ops.wm.quit_blender()
